Criando o processo de delete de tasks

+++ backend/src/main/java/com/br/api/backend/model/TaskModel.java
@@ -22,7 +22,7 @@ public class TaskModel {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long idTask;
+    private int idTask;
 
     private String nameTask;
